Botão de ajustes foi pra dentro do cabeçalho
