+++ src/Components/Cabecalho/style.tsx
@@ -2,8 +2,9 @@ import styled from "styled-components";
 import { global } from "../../globalStyle";
 
 export const CabecaDiv = styled.header `
-    display: inline-flex;
+    display: flex;
     align-items: center;
+    justify-content: space-between;
     padding-top: 4.5vh;
 
     width: 100vw;
@@ -37,17 +38,22 @@ export const CabecaDiv = styled.header `
 `;
 
 export const PageTitle = styled.h1 `
-    width: 100vw;
-    position: absolute;
-    left:0;
-
     font-size: 16px;
     font-family: ${global.fonts.quickSand};
     font-weight: 700;
-    text-align: center;
 
     color: ${global.colors.cinza};
 `; 
+
+export const DivAjustes = styled.div `
+    width: 6.66vw;
+    margin-right: 4.4vw;
+`;
+
+export const IconeAjustes = styled.img `
+    width: 6.66vw;
+    height: 6.66vw;
+`;
 
 export const BlurDiv = styled.div `
     display: none;
